+++ src/parser/companions/types/ConjureFey.mjs
@@ -44,7 +44,7 @@ export async function getConjureFey({
   const descriptionArray1 = raw.split("Fey creature of your choice.").pop().trim();
   const descriptionArray2 = descriptionArray1.split("</p>");
 
-  const description = `<p><em><strong>Psychic Attack.</em></strong>${descriptionArray2[0]}</p>`;
+  const description = `<p><em><strong>Psychic Attack.</strong></em> ${descriptionArray2[0]}</p>`;
 
   const manager = new DDBCompanionMixin(description, {}, { addMonsterEffects: true });
   manager.npc = stub;