+++ src/parser/enrichers/monster/FlamingSphere/_module.mjs
@@ -1,0 +1,2 @@
+export { default as FlameDamage } from "./FlameDamage.mjs";
+export { default as MoveAndAttack } from "./MoveAndAttack.mjs";