+++ src/parser/enrichers/monster/_module.mjs
@@ -5,3 +5,4 @@ export * as Dragon from "./Dragon/_module.mjs";
 export * as ConjuredAnimals from "./ConjuredAnimals/_module.mjs";
 export * as ConjuredElemental from "./ConjuredElemental/_module.mjs";
 export * as ConjuredFey from "./ConjuredFey/_module.mjs";
+export * as FlamingSphere from "./FlamingSphere/_module.mjs";